// generated/src/app/api-models/smdefauth-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmdefauthS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  defauthConsoleuser: string;
  defauthOsusername: string;
  defauthIsosusernameset: number;
  defauthOspassword: string;
  defauthIsospasswordset: number;
  defauthDbusername: string;
  defauthIsdbusernameset: number;
  defauthDbpassword: string;
  defauthIsdbpasswordset: number;
  defauthInternalpassword: string;
  defauthIsinternalpasswordset: number;
}